Automates PR creation on Valkey website for new release (#13)
This creates a pull request on valkey-io.github.io that will update the website with the newest release.

If the new release is a major, minor and patch version then a new file will be created. 
If the new release is an RC then this will update the file for version that corresponds to that RC.

---------
Signed-off-by: Nikhil Manglore <nmanglor@amazon.com>